creating main.py
